+++ mergesort.hpp
@@ -1,6 +1,6 @@
 #ifndef MERGESORT_HPP
 #define MERGESORT_HPP
 
-void mergesort(double* a);
+void mergesort(int* a);
 
 #endif // MERGESORT_HPP
